backend/requirements.txt
# K-IFRS 1019 DBO Validation System
# Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.3

# Data Validation
pydantic==2.5.3

# AI Integration
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dateutil==2.8.2

# Database (Supabase)
supabase>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0